修改了编码器的滤波滤去了十分大的尖峰和很小的个位数值，逐飞编码器返回的是定时器寄存器的值，所以放在轮询中会使得时间周期不统一而不是一个速度是位移

+++ Seekfree_TC264_Opensource_Library/CODE/Steer.h
@@ -15,6 +15,6 @@
 #define STEER_LEFT  850                 //舵机向左打死值
 #define STEER_RIGHT 680                 //舵机向右打死值
 
-void SteerCtrl(uint16 pwm);     //控制舵机
+void SteerCtrl(uint32 pwm);     //控制舵机
 
 #endif /* CODE_STEER_H_ */
